requirements.txt
# Core Framework
streamlit>=1.31.0

# Environment Management
python-dotenv>=1.0.0

# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LangChain Integrations
langchain-google-genai>=1.0.0
langchain-pinecone>=0.1.0
langchain-cohere>=0.1.0
langchain-text-splitters>=0.0.1

# Vector Database
pinecone-client>=3.0.0

# API Clients
google-generativeai>=0.3.0
cohere>=4.0.0
